Move ParachainRegistrar to primitives
Lift the trait out of pallet-registrar-relay so the
relay pallet and paras_registrar can share the registry
interface without depending on each other.
