+++ substrate/frame/registrar/primitives/Cargo.toml
@@ -18,6 +18,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { workspace = true, features = ["derive"] }
 scale-info = { workspace = true, features = ["derive"] }
 sp-core = { workspace = true }
+sp-runtime = { workspace = true }
 
 [features]
 default = ["std"]
@@ -25,4 +26,5 @@ std = [
 	"codec/std",
 	"scale-info/std",
 	"sp-core/std",
+	"sp-runtime/std",
 ]
